Fused attention, scan over encoder layers, fp16 with dynamic loss scaling
- _EncoderBlock now does QKV projection manually and calls
  jax.nn.dot_product_attention (fused SDPA, dispatches to cuDNN's fused
  kernel on GPU) instead of nn.MultiHeadDotProductAttention's separate
  matmul/softmax/matmul.
- The 12 encoder blocks are now nn.scan'd (one compiled loop body,
  stacked per-layer params) instead of Python-unrolled — reduces
  compiled graph size and host-dispatch overhead. Breaking change to
  the param pytree shape; no real checkpoints exist yet to migrate.
- compute_dtype default switched bfloat16 -> float16: T4 is Turing
  (sm_75), whose Tensor Cores are FP16-optimized — full-rate bf16
  Tensor Core support only arrived with Ampere, which likely explains
  why bf16 gave ~0 measured speedup despite the workload profiling as
  compute-bound (batch-size sweep showed ~flat images/s, not launch-
  overhead-bound as first suspected). fp16's narrow dynamic range needs
  loss scaling to avoid gradient underflow, unlike bf16 -- added via
  flax.training.dynamic_scale.DynamicScale on LeJEPATrainState, wired
  through both train_step (train_loop.py) and native_train_step
  (train_loop_native.py), including checkpoint save/load and eval.py's
  checkpoint-loading template.

Local jax/flax were stale (0.4.38/0.10.4) and hit a real
flax-internal incompatibility in nn.scan; upgraded to 0.11.0/0.12.8 to
match what Colab's `pip install -U` already installs there.

Test fixtures pin compute_dtype="float32" explicitly, since fp16
dot_general isn't supported on JAX's CPU backend at all (GPU-only by
nature) -- this repo has no GPU, so fp16 itself is unverified beyond
shape/API correctness on the fp32 code path.

+++ ljx_tests/test_lejepa.py
@@ -10,8 +10,9 @@ from ljx.models.lejepa import LeJEPAConfig, invariance_loss, lejepa_loss
 
 
 def tiny_config():
+    # float32: CPU dot_general doesn't support fp16 (the production default)
     return LeJEPAConfig(
-        backbone=ViTConfig(depth=2, embed_dim=24, num_heads=2),
+        backbone=ViTConfig(depth=2, embed_dim=24, num_heads=2, compute_dtype="float32"),
         projector_hidden_dims=(32, 32),
         projector_output_dim=16,
         sigreg=SigRegConfig(num_projections=32, seed=7),
